added fix for statemind-78 and removed unnecessary uint256 typecasting

+++ contracts/peer-to-peer/interfaces/ILenderVaultImpl.sol
@@ -65,7 +65,7 @@ interface ILenderVaultImpl {
     function updateLoanInfo(
         uint128 repayAmount,
         uint256 loanId,
-        uint256 collAmount,
+        uint128 collAmount,
         address collTokenCompartmentAddr,
         address collToken
     ) external;
